refactor: migrate ci_detect, adr_scanner, logical_location, tool_metadata to capability packages
(cherry picked from commit 9b816127c86d7697cbbfc513e220eae6c051ae1b)
